Extract quadNA from classes
I have added an Init funtion that will load the dll and quadNA the function, and set the kafka namespace in both the Producer and Consumer class.

I have removed the Consume and Produce functions and renamed Consume1 to Consume and Produce1 to Produce.
I have added Apl cover functions for all the calls to the dll that need it.